app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
asyncpg==0.29.0
redis==5.0.8
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
httpx==0.27.2
APScheduler==3.10.4
python-multipart==0.0.9
